Add setup.py prototype and rework requirements.txt
This commit adds a fairly minimal setup.py and clears up additional
requirements.

The plan is to make pybm have as few requirements as necessary
outside of the standard library to make it portable. This means that
the benchmarking process will require installing one's benchmarking
library of choice, or even write your own benchmarking class yourself.
All in all, this seems like an acceptable tradeoff for the user, who
gets a leaner package and more control, as long as there is a good
default benchmark class to use out of the box, preferably by an
optional pip dependency.