Add base command and command database to pybm
This commit implements the first command, the bare `pybm` command
with optional flags.

The first working flag is the version flag, which uses a special
argparse action construct detailed on the Python Argparse manpage.

Additionally, an unfrozen command database was added for storing
and retrieving commands through the main entrypoint. In the future,
this database (simple key-value store) should be locked to prevent
command injections.