Add `pybm create` command prototype
This commit adds a first prototype for the `pybm create` command.
It is designed to check out git worktrees and optionally install
Python virtual environments upon detection of overriding dependencies.

For now, only the git worktree creation is supposed to work. In
subsequent commits, better exception handling as well as the venv
building logic will be added.