Add commit SHA resolution, fix ref resolution bug detaching HEAD
This commit contains logic to facilitate git commit SHA resolution
for specified git SHA fragments in `pybm create`. Contrary to refs,
where you must use `git rev-list -n 1 <ref>`, if the object is already
a commit SHA fragment, `git rev-parse <ref>` is the right choice.

In addition, it fixes a bug where `git worktree add` failed to deal with
branch names in the format "refs/heads/$BRANCH" and would detach the
HEAD of the branch; this is fixed by not prepending the ref type
anymore.