Refactor git worktree wrapper class and add feature guard
This commit is a refactor of the git worktree wrapper class. Now
subprocess calls are dispatched dynamically depending on whether
output needs to be extracted (e.g. `git worktree list`) or just a
return code.
Also, a git feature guard was added, which is supposed to guard against
using git worktree features with an insufficient git version. It works
both on command level, as well as for single options.