Unify previous commands into new `pybm env` command
This commit is a major new orientation in pybm design. It departs
from single-shot commands `pybm create/destroy` a la Terraform and
instead formalizes the concept of `env` with the namesake command
to manage benchmark environments.

The spec of `pybm env` is tentatively:
* pybm env create
* pybm env delete
* pybm env list
* pybm env install
* pybm env update

This new concept should also make it easier to integrate other
Python package / environment management tools like poetry and pipenv
down the line.