Add Python environment builder class and stdlib implementation
This commit introduces the class interface for the Builder,
responsible for creating and managing Python virtual environments
tied to benchmark runs.

The implementation using `venv` from the Python standard library
resides in `pybm.builders.stdlib.PythonStdlibBuilder`, whereas the
abstract base interface is in the `builder` submodule in the same
source tree.

The class interface is pending some minor changes, it should in
general be considered stable now.