Add benchmark runner interface and stdlib and GBM implementations
This commit adds the benchmark runner class interface, which is
responsible for running benchmarks in a specified benchmark environment
and importing all of the symbols correctly.

Aside from the base class, which still has some changes and additional
functionality pending (most likely), there are two implementations
that are designed to work out of the box with pybm at launch:
* pybm.runners.stdlib, a benchmark runner based on timeit from the Python standard library,
* pybm.runners.gbm, a runner based on the Google Benchmark library,
or rather its Python bindings.

Since no working wheel CI exists for Google Benchmark as of this commit
(09/2021), users on a lot of platforms and Python versions will have to
build Google Benchmark wheels themselves, which requires Bazel. A working
alternative in the stdlib runner is most likely very welcome for a lot
of users that do not want to dive too deep into the details of Python pÃackaging.