Add `pybm init` command implementation
This commit adds an implementation of `pybm init`, designed to
initialize pybm in a git repository for benchmarking. Tests for
config file writing and environment discovery have been conducted,
and need to be formalized in the form of pytest targets.

Aside from the current status, future directions include:
* Specifying overriding options on the command line instead of
writing a completely default file and then changing it later
* Splitting off config options for special runner/builder/reporter
classes into subclasses of config groups and initializing the proper
groups based on class name