Add `pÃybm config` cmmand implementation
This commit adds a working implementation of `pybm config`, a command
designed to manipulate the pybm configuration file from the command line.
Under the hood, this process is facilitated by a special getter-setter mixin
class, which allows dotted attribute access for arbitrarily nested objects.

In total, `pybm config` is designed to have four subcommands:
* get, for displaying values or value groups
* set, for setting single option values
* list, for displaying all option groups
* describe, for printing further explanatory content regarding each option/group.

As of now, `pybm config describe` is not functional.