Add `pybm run` command implementation
This commit adds a working implementation of `pybm run`.

The run command probably has the most design uncertainty attached to
it out of all pybm commands. This is due to the blurred lines between
runner functionality and command functionality. As of now, functionality
has been split between the two in roughly the following way:

Ã- `pybm run` is responsible for option parsing, optional help/usage printing
and target discovery and validation. For each environment, it discovers its
targets based on a path relative to the corresponding git worktree.
- The associated runner is responsible for actually executing the benchmark
logic, which includes timing, timer probing to set the number of iterations,
optional collection of user counters, context providers, and saving the
results to files.

Some of the functionality is pending completion or change. Some more
command line switches will likely be added along the way.