Add pybm class specs and mixin util classes
This command formalizes some class objects integral to pybm, as well
as a few mixins for supplying functionality to a class.

Contents:
* Worktree, virtual environment specs describing a git worktree and
a python venv, respectively
* BenchmarkEnvironment, collection of all ingredients needed for a
reproducible, repeatable, versioned benchmark run
* Config groups, dataclasses holding a number of configuration values
for different components of the benchmarking pipeline

A downside is that the naming here is potentially confusing,
with the Python virtual environment actually being a subattribute of
the benchmark environment. A naming change would not be ideal, but
if a better, less confusing alternative comes up, not the end of the world,
either.