Refactor environment store to be a context manager
This commit re-introduces the environment store as a context manager,
which is a choice that has made a lot of the pybm command implementation
work more coherent.

In essence, the environment store tracks the state of environments in your
git repository by managing a YAML file as a resource, which is a textbook
use case for a context manager. With this approach, previously buggy
implementations of creation, modification and deletion of environments
was fixed, too.

The only major thing outstanding is the update command, which requires
more thought toward which parts (if not all) of a benchmark environment
should be immutable and which should not. Given previous experience in
state tracking and management, it might be wise to completely pull the
plug on all in-place modification except package installations (and,
possibly, virtual environment linkage, but that would just warrant
recreation of the env spec object).