Add pybm config object for stateful configuration management
This commit adds the `PybmConfig` object, which keeps track of
customizable settings throughout multiple stages of the pybm
benchmarking pipeline, allowing the user to customize his or her
benchmarking experience.

The implementation is very simple, a getter-setter mixin enables
dotted access, while the substages of the pipeline are separated into
so called contÂfiguration groups.

In the future, single groups may not have to hold all information for
all available runners, since that would scale the group object sizes
with the runners (and bloat config files unnecessarily with unused
information). Rather, subgroups in the form of subclasses of the
base groups should do the trick of siloing functionality for different
benchmark component implementations.