Change git wrapper to be a configurable command class member
This commit changes the GitWorktreeWrapper class to incorporate
state (read: load the global config) into its methods. At the moment,
it is limited to one configuration option (whether to create the worktree
in the parent directory or not), but this could expand in the future.

Alongside, some refactors to the environment store took place, incorporating
the changes to the benchmark environment spec.