Add context to pybm.run, value canonicalization for config setters
This commit introduces some changes for a few first-class pybm concepts:
* The `pybm.run` API is given the source file global context as a
keyword argument. This is necessary to discover the targets properly.
* The `set_value` API of the state mixin now canonicalizes values upon
setting. It looks at the annotation of the variable to set, and runs a
few checks against the input to assure that the set value makes sense.
This is especially important for booleans, which can now only be set
through the "true/false" literals (case-insensitive).