Fix critical venv creation symlink bug
This commit fixes a serious bug in virtual environment creation that
could lead to unintended destruction of the root environment as well.

The problem was that during venv creation with a subprocess, the default
mode is that symlinks to Python and pip executables are created. But if the
dispatching Python interpreter was itself a symlink (i.e. from a venv in an
IDE), a symlink to this other symlink would be introduced. Then, upon
destruction with shutil.rmtree, symlinks would be resolved, destroying a portion
of the environment. Resolving the creating symlink to the underlying system executable
solves the problem.

Other than that, a wrong root directory that could also lead to the wrong environment
being deleted was fixed as well.