Refactor env command and environment store
This commit is a refactor of benchmark environment handling. Now,
the lifecycle of benchmark environments is entirely at the discretion
of the EnvironmentStore class. The `pybm env` command is not a thin
wrapper around the environment store, previous functionality was moved
to the environment store.

Additionally, proper context management was introduced to the environment
store. It remains now to extend the context management workflow for print
statements in other operations.