Enable running benchmark targets as Python modules
Previously, the benchmark runner's `dispatch` method dispatched all
subprocesses with the `python benchmark_file.py [options]` syntax. This
unfortunately leads to a lot of trouble when benchmarking code that is not
contained in a package, i.e. cannot be discovered using `sys.path`. Instead
of requiring users to write a setup.py for each of their target packages,
this commit enables running the targets as modules instead.

To do this, the user simply uses the `-m` switch during `pybm run`, similaryly
to how he would invoke a file as a module with the standard Python interpreter.