Add reducer step to processing pipeline
This commit adds reducers in their simplest form. Right now, mean
and standard deviation are computed for numeric columns, and reported
for time values and user counters.

The `repetitions` command line argument is now properly supported on the timeit
benchmark runner.
The flag building API was changed on the Google Benchmark runner to accomodate the
latest changes from the module refactor.