Change from YAML to TOML for configuration files
This commit contains the move from YAML to TOML for all pybm configuration files.

All YAML calls have been substituted for their respective TOML calls. Mypy configuration was adapted to TOML, as was the readme note and requirements.txt file.

Due to TOML's object hierarchy, which prevents list-like structures, the benchmark environment container in the `EnvironmentStore` was changed from a list to a dict.
All discovery methods, lifecycle management and other APIs were adapted to this circumstance.

Previously hardcoded config paths across the project were changed to a single constant embedded in the `pybm.config` module, which is also the default value for the `PybmConfig.load()` API.