Add global config discovery via git, fix version parsing for Windows
This commit enables pybm benchmark runs on Windows systems by fixing the git version tuple parsing from the output of `git version`. Previously, parsing would fail due to a trailing dot in the string match
obtained from the regex matching, which happened because git versions on Windows are of the form MAJ.MIN.PATCH.windows.X. With the previous regex, the dot after PATCH would be picked up, leading to an
empty string in the string split, causing it to fail. The new regex picks up groups of dots AND THEN numbers, ensuring that a capture group never ends on a dot.

Additionally, a method to find the local config in the main worktree with git was added, which previously broke benchmarking in repositories with the pybm config ignored from source control.