Simplify TimeitRunner execution
This commit simplifies the TimeitRunner class. A loop was removed by passing the repetitions argument to the Timer.repeat API of timeit, which then returns a list that can be distributed into multiple JSON
results. At the end, benchmark results are assembled as a shallow copy of a local benchmark template constructed at the beginning of the run.