Add generalized IO operation, expose through the reporter class
This commit adds a generalized IO component for pybm, which allows to save benchmark results obtained from subprocesses in different ways. As a first IO class, the JSONFileIO was added, which reads and
writes JSON files locally to/from disk.

The ConsoleReporter was greatly simplified as a result. The main benchmark result processing logic is now much shorter and more concise. Formatting, reducing and comparing was moved to external APIs
specific to the ConsoleReporter.