Rename builder to provider component
This commit represents a name change from "builder" components to "provider" components. The latter is a tad more appropriate and known in the community for environment management (think terraform).

The only currently implemented provider is still venv via Python. This change removes the ties to Python a little bit, which could prove beneficial when extending git-based benchmarking to other languages in the future.