Change `env` concept name to `workspace`, materialize config directly
This commit changes the concept of the benchmark environment to the name
workspace. The workspace now entails both git and Python configuration
information, and inherits some APIs from the worktree and PythonSpec
abstractions.

Also, the PybmConfig class is now materialized inside the `pybm.config`
module directly, which makes a lot of config import statements much
easier and prevents the need to materialize the config every time in a
source file. The last remaining file with hands-on config management is
the implementation of `pybm init`.