Add error handling to timeit runner, refactor some utilities
This commit introduces error handling to the timeit runner, modeled
after the error handling in Google Benchmark.

Concretely, Google Benchmark does not kill the process when encountering
an error, but rather adds the `error_occurred` field (with the boolean value
"true") and the "error_message" field containing the error message as a
string. The same design was chosen for the timeit runner, and
implemented in this commit.

Furthermore, upon error, all timings are set to zero, which also mirrors
GBM's behavior.